requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
Flask-Cors==4.0.0
flower==2.0.1
humanize==4.8.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.1
MarkupPy==1.14
MarkupSafe==2.1.3
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
threadpoolctl==3.2.0
tornado==6.3.3
typing-extensions==4.7.1
tzdata==2023.3
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.16.2
